Refactor processing and fix caching and fix tests
