+++ src/Reports/index.ts
@@ -6,22 +6,11 @@ import {
   trancheSnapshotsPostProcess,
   trancheSnapshotsQuery,
 } from '../queries/trancheSnapshots.js'
-import { processBalanceSheetData } from './processors/balanceSheet.js'
-import { combineLatest, of } from 'rxjs'
-import { ReportProcessorType } from './processors/index.js'
+import { combineLatest, defer } from 'rxjs'
+import { processor } from './Processor.js'
 
 import { map } from 'rxjs'
-import { GroupBy } from '../utils/date.js'
-export interface ReportFilter {
-  from?: string
-  to?: string
-  groupBy?: GroupBy
-}
-
-export interface ReportData {
-  timestamp: string
-  [key: string]: unknown
-}
+import { ReportFilter } from './types.js'
 
 export class Reports extends Entity {
   constructor(
@@ -32,40 +21,41 @@ export class Reports extends Entity {
   }
 
   balanceSheet(filter?: ReportFilter) {
-    return this.generateReport('balanceSheet', filter)
+    return this._generateReport('balanceSheet', filter)
   }
 
-  generateReport(type: ReportProcessorType, filter?: ReportFilter) {
+  _generateReport(type: 'balanceSheet', filter?: ReportFilter) {
     return this._root._query(
-      [type, ...(filter ? [filter] : [])],
-      () => {
-        const dateFilter = {
-          timestamp: {
-            greaterThan: filter?.from,
-            lessThan: filter?.to,
-          },
-        }
+      [type, filter?.from, filter?.to, filter?.groupBy],
+      () =>
+        defer(() => {
+          const dateFilter = {
+            timestamp: {
+              greaterThan: filter?.from,
+              lessThan: filter?.to,
+            },
+          }
 
-        return combineLatest([
-          this.poolSnapshots({
+          const poolSnapshots$ = this.poolSnapshots({
             ...dateFilter,
             poolId: { equalTo: this.poolId },
-          }),
-          this.trancheSnapshots({
+          })
+          const trancheSnapshots$ = this.trancheSnapshots({
             ...dateFilter,
             tranche: { poolId: { equalTo: this.poolId } },
-          }),
-        ]).pipe(
-          map(([poolSnapshots, trancheSnapshots]) => {
-            switch (type) {
-              case 'balanceSheet':
-                return processBalanceSheetData({ poolSnapshots, trancheSnapshots }, filter)
-              default:
-                throw new Error(`Unsupported report type: ${type}`)
-            }
           })
-        )
-      },
+
+          return combineLatest([poolSnapshots$, trancheSnapshots$]).pipe(
+            map(([poolSnapshots, trancheSnapshots]) => {
+              switch (type) {
+                case 'balanceSheet':
+                  return processor.balanceSheet({ poolSnapshots, trancheSnapshots }, filter)
+                default:
+                  throw new Error(`Unsupported report type: ${type}`)
+              }
+            })
+          )
+        }),
       {
         valueCacheTime: 120,
       }
@@ -78,9 +68,5 @@ export class Reports extends Entity {
 
   trancheSnapshots(filter?: TrancheSnapshotFilter) {
     return this._root._queryIndexer(trancheSnapshotsQuery, { filter }, trancheSnapshotsPostProcess)
-  }
-
-  _processBalanceSheet(data: any, filter?: ReportFilter) {
-    return processBalanceSheetData(data, filter)
   }
 }
