--- src/Reports/processors/index.ts
@@ -1,7 +1,0 @@
-import { processBalanceSheetData } from './balanceSheet.js'
-
-export const processors = {
-  balanceSheet: processBalanceSheetData,
-} as const
-
-export type ReportProcessorType = keyof typeof processors
