+++ src/Reports/types.ts
@@ -1,3 +1,7 @@
+import { PoolSnapshot } from '../queries/poolSnapshots.js'
+import { TrancheSnapshot } from '../queries/trancheSnapshots.js'
+import { Price } from '../utils/BigInt.js'
+import { Currency } from '../utils/BigInt.js'
 import { GroupBy } from '../utils/date.js'
 
 export interface ReportFilter {
@@ -9,4 +13,27 @@ export interface ReportFilter {
 export interface ReportData {
   timestamp: string
   [key: string]: unknown
+}
+
+export interface BalanceSheetReport extends ReportData {
+  date: string
+  assetValuation: Currency
+  onchainReserve: Currency
+  offchainCash: Currency
+  accruedFees: Currency
+  netAssetValue: Currency
+  tranches?: {
+    name: string
+    timestamp: string
+    tokenId: string
+    tokenSupply: Currency
+    tokenPrice: Price | null
+    trancheValue: Currency
+  }[]
+  totalCapital?: Currency
+}
+
+export type BalanceSheetData = {
+  poolSnapshots: PoolSnapshot[]
+  trancheSnapshots: TrancheSnapshot[]
 }
